socket_client/requirements.txt
requests[socks]==2.24.0
websocket-client==0.57.0